+++ example/todomvc/src/modules/todos/actions/index.ts
@@ -1,0 +1,6 @@
+import { ActionReturnType } from 'react-fluxrx';
+
+export type actionType = ActionReturnType<typeof import('./index')>;
+
+export * from './items';
+export * from './visibility';