+++ example/todomvc/src/modules/todos/reducers/items.ts
@@ -8,15 +8,15 @@ export type stateType = Array<{
 
 const initialState: stateType = [
   {
+    id: 0,
     text: 'Use react-fluxRx',
     completed: false,
-    id: 0,
   },
 ];
 
 export const reducer = (state = initialState, action: actionType): stateType => {
   switch (action.type) {
-    case 'todos/TODO_ADD':
+    case 'TODOS/TODO_ADD':
       return [
         ...state,
         {
@@ -26,16 +26,16 @@ export const reducer = (state = initialState, action: actionType): stateType => 
         },
       ];
 
-    case 'todos/TODO_DELETE':
+    case 'TODOS/TODO_DELETE':
       return state.filter((todo) => todo.id !== action.payload.id);
 
-    case 'todos/TODO_EDIT':
+    case 'TODOS/TODO_EDIT':
       return state.map((todo) => (todo.id === action.payload.id ? { ...todo, text: action.payload.text } : todo));
 
-    case 'todos/COMPLETE_TODO':
+    case 'TODOS/COMPLETE_TODO':
       return state.map((todo) => (todo.id === action.payload.id ? { ...todo, completed: !todo.completed } : todo));
 
-    case 'todos/COMPLETE_ALL':
+    case 'TODOS/COMPLETE_ALL':
       const areAllMarked = state.every((todo) => todo.completed);
 
       return state.map((todo) => ({
@@ -43,7 +43,7 @@ export const reducer = (state = initialState, action: actionType): stateType => 
         completed: !areAllMarked,
       }));
 
-    case 'todos/COMPLETE_CLEAR_COMPLETED':
+    case 'TODOS/COMPLETE_CLEAR_COMPLETED':
       return state.filter((todo) => todo.completed === false);
 
     default: