+++ example/todomvc/src/store/reducer.ts
@@ -1,11 +1,9 @@
 import { combineReducers } from 'react-fluxrx';
 
-import todos from './todos/reducers';
+import todos from '../modules/todos/reducers';
 
 export const reducer = combineReducers({
   todos,
 });
 
 export default reducer;
-
-export type stateType = ReturnType<typeof reducer>;