+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 - Remove the bindActions function
 - Conversion to functional components and hooks in the example [TodoMVC with AJAX](./example/todomvc-ajax)
+- Conversion to functional components and hooks in the example [TodoMVC](./example/todomvc)
 
 ### Performance